sched: widen sysctl_sched_features storage to unsigned long

+++ kernel/sched/core.c
@@ -55,7 +55,7 @@ EXPORT_SYMBOL_GPL(runqueues);
  */
 #define SCHED_FEAT(name, enabled)	\
 	(1UL << __SCHED_FEAT_##name) * enabled |
-const_debug unsigned int sysctl_sched_features =
+const_debug unsigned long sysctl_sched_features =
 #include "features.h"
 	0;
 #undef SCHED_FEAT
